# Improve fnv64a hashing algorithm by remove all allocations