@janisz 2016-06-05
Hi @mateuszdyminski 

Thanks for PR. I left comments on it.