@mateuszdyminski 2016-06-06
Moved implementation to fnv.go file.

Created test and benchmark for new fnv hash implementation. Here is the output from benchmark:
> BenchmarkFnvHashSum64-4             100000000       11.1 ns/op     0 B/op         0 allocs/op
> BenchmarkFnvHashStdLibSum64-4   20000000         67.3 ns/op   16 B/op         2 allocs/op

I also removed all ReportAllocs(). Thanks for tip with -benchmem flag.