@janisz 2016-06-06
LGTM!
@druminski would you like to review this patch?