@druminski 2016-06-06
@mateuszdyminski @janisz I have just analyzed this PR and I like it very much, great job and thanks for the contribution :)