@janisz 2016-06-06
We need to generate new benchmarks results for [README](https://github.com/allegro/bigcache#benchmarks)