@druminski 2016-06-07
Ok, new bench results generated and updated in [readme](https://github.com/allegro/bigcache#benchmarks)