@mateuszdyminski 2016-06-07
Nice! Thanks for accepting PR : )